+++ packages/api-derive/src/alliance/index.ts
@@ -7,12 +7,12 @@ import { hasProposals as collectiveHasProposals, members as collectiveMembers, p
 export type { u32 } from '@polkadot/types';
 export type { AccountId } from '@polkadot/types/interfaces';
 
-export const members = collectiveMembers('allianceMotion');
+export const members = /*#__PURE__*/ collectiveMembers('allianceMotion');
 
-export const hasProposals = collectiveHasProposals('allianceMotion');
-export const proposal = collectiveProposal('allianceMotion');
-export const proposalCount = collectiveProposalCount('allianceMotion');
-export const proposalHashes = collectiveProposalHashes('allianceMotion');
-export const proposals = collectiveProposals('allianceMotion');
+export const hasProposals = /*#__PURE__*/ collectiveHasProposals('allianceMotion');
+export const proposal = /*#__PURE__*/ collectiveProposal('allianceMotion');
+export const proposalCount = /*#__PURE__*/ collectiveProposalCount('allianceMotion');
+export const proposalHashes = /*#__PURE__*/ collectiveProposalHashes('allianceMotion');
+export const proposals = /*#__PURE__*/ collectiveProposals('allianceMotion');
 
-export const prime = collectivePrime('allianceMotion');
+export const prime = /*#__PURE__*/ collectivePrime('allianceMotion');