+++ packages/api-derive/src/staking/stakerExposure.ts
@@ -46,7 +46,7 @@ export function stakerExposures (instanceId: string, api: DeriveApi): (accountId
   );
 }
 
-export const stakerExposure = firstMemo(
+export const stakerExposure = /*#__PURE__*/ firstMemo(
   (api: DeriveApi, accountId: Uint8Array | string, withActive?: boolean) =>
     api.derive.staking.stakerExposures([accountId], withActive)
 );